+++ CHANGELOG.md
@@ -7,14 +7,21 @@
 - improve the code and Verify the test cases
 - Improved method handling for multi-record operations instead of self
 
-## [18.0.1.3.0] - 2025-04-03 | Borrow Transaction History
+## [18.0.1.3.1] - 2025-04-03 | Borrow Transaction History
 
 - Change Field and button names
 - Enhanced model-level doc string
-- Fixed dependency error 
+- Fixed dependency error
+- Fixed on hand quantity issue (decrease in all condition)
+- Make require deposit amount field using constrains
+- Added search view
 
 ## [18.0.1.4.0] - 2025-04-03 | Schedule Actions
 
 - Crete method for handling actions
 - Update the reminder string to dynamic
 - create ir_action and cron action
+
+## [18.0.1.5.0] - 2025-04-03 | Mail template
+
+-Create mail template